# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Church Projection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List all songs
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: collectionId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Songs list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"
    post:
      operationId: createSong
      tags: [songs]
      summary: Create a song
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSongBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
  /songs/{id}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get a song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSong
      tags: [songs]
      summary: Update a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSongBody"
      responses:
        "200":
          description: Updated song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /songs/{id}/verses:
    get:
      operationId: getSongVerses
      tags: [songs]
      summary: Get parsed verses of a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SongVerse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Song:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        lyrics:
          type: string
        mp3Path:
          type: string
          nullable: true
        key:
          type: string
          nullable: true
        bpm:
          type: integer
          nullable: true
        category:
          type: string
          nullable: true
        collectionId:
          type: integer
          nullable: true
        collectionName:
          type: string
          nullable: true
        coverImage:
          type: string
          nullable: true
        isFavorite:
          type: boolean
        createdAt:
          type: string
      required: [id, title, author, lyrics, isFavorite, createdAt]
    SongVerse:
      type: object
      properties:
        index:
          type: integer
        label:
          type: string
        lines:
          type: array
          items:
            type: string
      required: [index, label, lines]
    CreateSongBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        lyrics:
          type: string
        mp3Path:
          type: string
        key:
          type: string
        bpm:
          type: integer
        category:
          type: string
        collectionId:
          type: integer
        coverImage:
          type: string
      required: [title, author, lyrics]
    UpdateSongBody:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        lyrics:
          type: string
        mp3Path:
          type: string
        key:
          type: string
        bpm:
          type: integer
        category:
          type: string
        collectionId:
          type: integer
        coverImage:
          type: string
        isFavorite:
          type: boolean
...